Backend Foreign Key & routes

+++ backend/controllers/comment.js
@@ -1,11 +1,11 @@
 "use strict"
 
-const Comment = require('../models/commentModel')
+const { Comment } = require('../models/')
 
 exports.createComment = (req, res, next) => {
   Comment.create({
     description: req.body.description,
-    user_id: req.body.user_id,
+    userId: req.body.userId,
     postId: req.body.postId
   })
     .then(() => res.status(201).json({ message: 'Commentaire créé !' }))
@@ -21,9 +21,11 @@ exports.getAllComments = (req, res, next) => {
 exports.deleteComment = (req, res, next) => {
   Comment.findOne({ id: req.params.commentId })
     .then(Comment => {
-      Comment.destroy({ where: { id: req.params.commentId } })
-        .then(() => res.status(200).json({ message: 'Commentaire supprimé !' }))
-        .catch(error => res.status(400).json({ error }))
+      if (req.userAuth.id == req.params.id || req.userAuth.isAdmin) {
+        Comment.destroy({ where: { id: req.params.commentId } })
+          .then(() => res.status(200).json({ message: 'Commentaire supprimé !' }))
+          .catch(error => res.status(400).json({ error }))
+      } else return res.status(401).json({ error: 'Action non autorisée !' })
     })
     .catch(error => res.status(500).json({ error }))
 }
