+++ backend/controllers/post.js
@@ -1,6 +1,6 @@
 "use strict"
 
-const Post = require('../models/postModel')
+const { Post } = require('../models')
 const fs = require('fs')
 
 exports.createPost = (req, res, next) => {
@@ -27,13 +27,15 @@ exports.getOnePost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   Post.findOne({ where: { id: req.params.id } })
     .then(post => {
-      // On extrait le nom du fichier à supprimer
-      const filename = post.imageUrl.split('/images/')[1]
-      fs.unlink(`images/${filename}`, () => {
-        Post.destroy({ where: { id: req.params.id } })
-          .then(() => res.status(200).json({ message: 'Post supprimé !' }))
-          .catch(error => res.status(400).json({ error }))
-      })
+      if (req.userAuth.id == req.params.id || req.userAuth.isAdmin) {
+        // On extrait le nom du fichier à supprimer
+        const filename = post.imageUrl.split('/images/')[1]
+        fs.unlink(`images/${filename}`, () => {
+          Post.destroy({ where: { id: req.params.id } })
+            .then(() => res.status(200).json({ message: 'Post supprimé !' }))
+            .catch(error => res.status(400).json({ error }))
+        })
+      } else return res.status(401).json({ error: 'Action non autorisée !' })
     })
     .catch(error => res.status(500).json({ error: JSON.stringify(error) }))
 }
